+++ src/super/i-block/modules/async-render/index.ts
@@ -204,13 +204,7 @@ export default class AsyncRender extends Friend {
 									})
 
 									.catch((err) => {
-										const
-											{methods} = this.meta;
-
-										if (methods.errorCaptured) {
-											methods.errorCaptured.fn.call(this.component, err);
-										}
-
+										stderr(err);
 										return breaker;
 									})
 							};
@@ -259,6 +253,13 @@ export default class AsyncRender extends Friend {
 			const
 				{async: $a} = this;
 
+			let
+				group = 'asyncComponents';
+
+			if (opts.group != null) {
+				group = `asyncComponents:${opts.group}:${chunkI}`;
+			}
+
 			for (let o = newIterator, el = o.next(); !el.done; el = o.next()) {
 				let
 					val = el.value;
@@ -269,20 +270,18 @@ export default class AsyncRender extends Friend {
 				if (isValPromise) {
 					try {
 						// eslint-disable-next-line require-atomic-updates
-						val = await val;
+						val = await $a.promise(<Promise<unknown>>val, {group});
 
 						if (val === breaker) {
 							break;
 						}
 
 					} catch (err) {
-						const
-							{methods} = this.meta;
-
-						if (methods.errorCaptured) {
-							methods.errorCaptured.fn.call(this.component, err);
+						if (err?.type === 'clearAsync') {
+							break;
 						}
 
+						stderr(err);
 						continue;
 					}
 				}
@@ -300,18 +299,13 @@ export default class AsyncRender extends Friend {
 						Object.isArray(iterable) && total >= iterable.length;
 
 					if (needRender) {
-						const
-							desc = <TaskDesc>{};
-
-						let
-							group = 'asyncComponents';
+						const desc: TaskDesc = {
+							renderGroup: group
+						};
 
 						if (opts.group != null) {
-							group = `asyncComponents:${opts.group}:${chunkI}`;
 							desc.destructor = () => $a.terminateWorker({group});
 						}
-
-						desc.renderGroup = group;
 
 						const
 							els = <Node[]>cb(newArray, desc);
@@ -391,8 +385,8 @@ export default class AsyncRender extends Friend {
 
 				return exec(res);
 
-				function exec(res: unknown): boolean {
-					if (Object.isTruly(res)) {
+				function exec(passed: unknown): boolean {
+					if (Object.isTruly(passed)) {
 						cb();
 						return true;
 					}