+++ CHANGELOG.md
@@ -20,6 +20,9 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 #### :bug: Bug Fix
 
 * Fixed a bug when `iterate` takes the rejected promise `i-block/modules/async-render`
+* Fixed a bug with redundant `v-for` invokes:
+  * `i-block/modules/async-render`
+  * `core/component/render-function`
 
 ## v3.0.0-rc.104 (2020-12-07)